backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
playwright==1.40.0
validators==0.22.0
python-dotenv==1.0.0
google-generativeai==0.3.0

